Client and server can take command-line arguments containing port number
If you find running the server fails with "port already in use", someone else has started a server on turing already. Don't worry, you can start another, you just need a different "port number". Think of a number between 50000 and 59999 (it doesn't matter what it is so long as nobody's using it) and use it in place of 51234 in the following:

gradle server:run --args="51234"
gradle client:run --args="51234"